+++ src/main/java/ee/carlrobert/llm/client/DeserializationUtil.java
@@ -2,6 +2,7 @@ package ee.carlrobert.llm.client;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import ee.carlrobert.llm.client.codegpt.response.CodeGPTException;
 import java.io.IOException;
 import okhttp3.Response;
 
@@ -15,10 +16,20 @@ public class DeserializationUtil {
 
   public static <T> T mapResponse(Response response, Class<T> clazz) {
     var body = response.body();
+    if (body == null) {
+      throw new RuntimeException("Response body is null");
+    }
+
+    String json = "";
     try {
-      return OBJECT_MAPPER.readValue(body.string(), clazz);
+      json = body.string();
+      return OBJECT_MAPPER.readValue(json, clazz);
     } catch (IOException ex) {
-      throw new RuntimeException("Could not deserialize response", ex);
+      try {
+        throw OBJECT_MAPPER.readValue(json, CodeGPTException.class);
+      } catch (IOException e) {
+        throw new RuntimeException("Could not deserialize response", ex);
+      }
     }
   }
 }