+++ src/main/java/ee/carlrobert/llm/client/codegpt/request/prediction/PastePredictionRequest.java
@@ -1,0 +1,14 @@
+package ee.carlrobert.llm.client.codegpt.request.prediction;
+
+public class PastePredictionRequest extends PredictionRequest {
+
+  private final String pastedCode;
+
+  public PastePredictionRequest(String pastedCode) {
+    this.pastedCode = pastedCode;
+  }
+
+  public String getPastedCode() {
+    return pastedCode;
+  }
+}