+++ src/main/java/ee/carlrobert/llm/client/codegpt/response/PredictionResponse.java
@@ -1,0 +1,14 @@
+package ee.carlrobert.llm.client.codegpt.response;
+
+public class PredictionResponse {
+
+  private String nextRevision;
+
+  public String getNextRevision() {
+    return nextRevision;
+  }
+
+  public void setNextRevision(String nextRevision) {
+    this.nextRevision = nextRevision;
+  }
+}