+++ src/main/java/ee/carlrobert/llm/client/openai/completion/request/Prediction.java
@@ -1,0 +1,20 @@
+package ee.carlrobert.llm.client.openai.completion.request;
+
+public class Prediction {
+
+  private final String type;
+  private final String content;
+
+  public Prediction(String type, String content) {
+    this.type = type;
+    this.content = content;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public String getContent() {
+    return content;
+  }
+}